# projects/neural-sync/artifacts/api-contracts.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEURAL SYNC API",
    "version": "1.0.0",
    "description": "Multi-dimensional AI-driven developer–project compatibility engine. All protected routes require JWT Bearer authentication per ADR-002. The 500 ms p95 latency SLA applies to the synchronous POST /matches response per ADR-003. LLM explanations are generated asynchronously (Phase-1: Google Gemini; Claude API is a supported alternate per ADR-004); the initial response always contains a deterministic stub (≥50 chars) per ADR-003/ADR-004. Error envelope on all 4xx/5xx: {error_code, message, request_id}.",
    "contact": {
      "name": "NEURAL SYNC Engineering",
      "email": "engineering@neural-sync.example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development server"
    },
    {
      "url": "https://neural-sync.example.com/api/v1",
      "description": "Production server"
    }
  ],
  "security": [
    { "BearerAuth": [] }
  ],
...
  "paths": {
...
    "/projects": {
      "post": {
        "tags": ["projects"],
        "summary": "Ingest a new project profile",
        "description": "Creates a ProjectProfile. Validates workload_intensity ∈ [0.0, 1.0] and innovation_level ∈ [0.0, 1.0]. Enqueues skill embedding generation as a background task. Manager role required.",
        "operationId": "createProject",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectProfileCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project profile created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectProfile" }
              }
            }
          },
          "400": {
            "description": "Invalid profile data",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden — manager role required to create projects",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "409": {
            "description": "Project with provided id already exists",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "422": {
            "description": "Validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/projects/{project_id}": {
      "get": {
        "tags": ["projects"],
        "summary": "Get a project profile by ID",
        "operationId": "getProject",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectIdPath" }
        ],
        "responses": {
          "200": {
            "description": "Project profile found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectProfile" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Project not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      },
      "put": {
        "tags": ["projects"],
        "summary": "Replace a project profile",
        "description": "Full replacement of a ProjectProfile. Re-enqueues embedding generation on success. Manager role required.",
        "operationId": "updateProject",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectIdPath" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectProfileCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project profile updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectProfile" }
              }
            }
          },
          "400": {
            "description": "Invalid profile data",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Project not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "422": {
            "description": "Validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      },
      "delete": {
        "tags": ["projects"],
        "summary": "Delete a project profile",
        "description": "Deletes a ProjectProfile and its embeddings. Historical MatchRecord rows are retained for analytics. Manager role required.",
        "operationId": "deleteProject",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectIdPath" }
        ],
        "responses": {
          "204": {
            "description": "Project profile deleted. No response body."
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden — manager role required",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Project not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from POST /auth/login. Algorithm: HS256. Payload: {sub: user_id, role: 'developer'|'manager', iat, exp}. TTL: 1h. Secret loaded from env NEURAL_SYNC_JWT_SECRET. Per ADR-002: store in memory (not localStorage) on the React frontend; refresh token stored in HttpOnly cookie."
      }
    },
    "parameters": {
...
      "ProjectIdPath": {
        "name": "project_id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" },
        "description": "Project unique identifier (UUID v4)"
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error envelope returned on all 4xx and 5xx responses.",
        "required": ["error_code", "message", "request_id"],
        "properties": {
          "error_code": {
            "type": "string",
            "description": "Machine-readable error code. Examples: NOT_FOUND, INVALID_WEIGHTS, VALIDATION_ERROR, ERASURE_FAILED, SERVICE_UNAVAILABLE, FORBIDDEN, UNAUTHORIZED, CONFLICT"
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the error, safe to display in developer tooling"
          },
          "request_id": {
            "type": "string",
            "description": "Unique request trace identifier for correlating logs"
          }
        }
      },
...
      "ProjectProfileCreate": {
        "type": "object",
        "description": "Request body for creating or replacing a ProjectProfile.",
        "required": [
          "required_skills",
          "team_structure",
          "workload_intensity",
          "innovation_level",
          "timezone_overlap_required",
          "duration_weeks",
          "growth_opportunities"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Optional client-provided UUID. Server generates one if omitted."
          },
          "name": {
            "type": "string",
            "description": "Human-readable project name"
          },
          "required_skills": {
            "type": "array",
            "items": { "type": "string" },
            "minItems": 1,
            "description": "Required technical skills for the project"
          },
          "team_structure": {
            "type": "string",
            "description": "Description of team composition and collaboration model, e.g. 'Cross-functional agile squad of 6, async-first, weekly sync'"
          },
          "workload_intensity": {
            "type": "number",
            "minimum": 0.0,
            "maximum": 1.0,
            "description": "Workload intensity 0.0 (low-pressure) to 1.0 (very high intensity). Values ≥ 0.8 over extended periods contribute to burnout risk."
          },
          "innovation_level": {
            "type": "number",
            "minimum": 0.0,
            "maximum": 1.0,
            "description": "0.0 = pure maintenance/support; 1.0 = cutting-edge R&D/greenfield"
          },
          "timezone_overlap_required": {
            "type": "string",
            "description": "Required timezone overlap window, e.g. 'UTC+0 to UTC+3', 'Americas (UTC-8 to UTC-5)'"
          },
          "duration_weeks": {
            "type": "integer",
            "minimum": 1,
            "description": "Projected project duration in weeks"
          },
          "growth_opportunities": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Learning and career growth opportunities the project offers, e.g. ['Kubernetes', 'Team leadership', 'ML pipeline design']"
          }
        }
      },
      "ProjectProfile": {
        "description": "Complete project profile including server-generated fields.",
        "allOf": [
          { "$ref": "#/components/schemas/ProjectProfileCreate" },
          {
            "type": "object",
            "required": ["id", "created_at", "updated_at"],
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "created_at": { "type": "string", "format": "date-time" },
              "updated_at": { "type": "string", "format": "date-time" }
            }
          }
        ]
      },
...
    }
  }
}